+++ src/com/exedio/cope/console/CopyConstraintCop.java
@@ -88,7 +88,7 @@ final class CopyConstraintCop extends TestCop<CopyConstraint>
 
 	@Override
 	@SuppressFBWarnings("NP_LOAD_OF_KNOWN_NULL_VALUE") // OK: caused by try-with-resources
-	int check(final CopyConstraint constraint, final Model model)
+	long check(final CopyConstraint constraint, final Model model)
 	{
 		try(TransactionTry tx = model.startTransactionTry("Console CopyConstraint " + id))
 		{
